azurekeyvault/azure-hosted-python/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
azure-keyvault-secrets==4.0.0
azure-identity==1.0.1
